fix: use draft/commit-on-blur pattern for custom format input
The custom worktree name format input was calling updateSettings on every
keystroke, which triggers prepareLocalWorktreeRootsForRepos repeatedly —
the same class of problem the workspaceDir field avoids with a draft
pattern. Reuse that pattern: edit a local draft, commit on blur/Enter,
reset on Escape.

Also extract WorktreeNamingControl into its own file to stay under the
400-line max-lines limit on WorkspaceDirectorySetting.tsx.
