+++ src/renderer/src/components/settings/WorkspaceDirectorySetting.tsx
@@ -13,7 +13,7 @@ import { Input } from '../ui/input'
 import { Label } from '../ui/label'
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '../ui/select'
 import { SearchableSetting } from './SearchableSetting'
-import { SettingsRow } from './SettingsFormControls'
+import { WorktreeNamingControl } from './WorktreeNamingControl'
 import { useSidebarHostScopeOptions } from '../sidebar/use-sidebar-host-scope-options'
 import {
   buildHostScopeChoices,
@@ -278,157 +278,4 @@ export function WorkspaceDirectorySetting({
 function isComposingKeyboardEvent(event: React.KeyboardEvent<HTMLInputElement>): boolean {
   const nativeEvent = event.nativeEvent
   return nativeEvent.isComposing || nativeEvent.keyCode === 229
-}
-
-type WorktreeNamingMode = 'flat' | 'nested' | 'custom'
-
-type WorktreeNamingControlProps = {
-  settings: GlobalSettings
-  updateSettings: (updates: Partial<GlobalSettings>) => void
-}
-
-// Why: the naming mode dropdown is the single UI control for worktree folder
-// layout. Each option writes the full set of related fields in one update so
-// the runtime, persistence history, and legacy readers all stay in sync.
-function WorktreeNamingControl({
-  settings,
-  updateSettings
-}: WorktreeNamingControlProps): React.JSX.Element {
-  const inputId = useId()
-  const mode: WorktreeNamingMode = settings.worktreeNamingMode ?? 'nested'
-  const customFormat = settings.worktreeNameFormat ?? ''
-
-  const selectMode = (next: WorktreeNamingMode): void => {
-    if (next === 'flat') {
-      updateSettings({
-        worktreeNamingMode: 'flat',
-        nestWorkspaces: false,
-        worktreeNameFormat: undefined
-      })
-      return
-    }
-    if (next === 'nested') {
-      // Why: nested is a format preset so the runtime uses one code path;
-      // nestWorkspaces stays true for back-compat with legacy readers.
-      updateSettings({
-        worktreeNamingMode: 'nested',
-        nestWorkspaces: true,
-        worktreeNameFormat: '{repoName}/{name}'
-      })
-      return
-    }
-    // custom: keep whatever format the user had, or default to a flat example
-    // so the input is not empty on first switch.
-    const seed =
-      customFormat && customFormat !== '{repoName}/{name}' ? customFormat : '{repoName}.{name}'
-    updateSettings({
-      worktreeNamingMode: 'custom',
-      nestWorkspaces: false,
-      worktreeNameFormat: seed
-    })
-  }
-
-  const setCustomFormat = (value: string): void => {
-    updateSettings({ worktreeNameFormat: value || undefined })
-  }
-
-  const example = renderWorktreeNameExample(mode, customFormat, settings.workspaceDir)
-
-  return (
-    <div className="space-y-2">
-      <SettingsRow
-        label={translate(
-          'auto.components.settings.WorkspaceDirectorySetting.7a8b9c0dae',
-          'Worktree Naming'
-        )}
-        description={
-          <>
-            <span>
-              {mode === 'flat' &&
-                translate(
-                  'auto.components.settings.WorkspaceDirectorySetting.c3d4e5f6ab',
-                  'Worktrees are created directly under the workspace directory.'
-                )}
-              {mode === 'nested' &&
-                translate(
-                  'auto.components.settings.WorkspaceDirectorySetting.d4e5f6a7bc',
-                  'Worktrees are grouped under a repo-named subfolder.'
-                )}
-              {mode === 'custom' &&
-                translate(
-                  'auto.components.settings.WorkspaceDirectorySetting.e5f6a7b8cd',
-                  'Use {repoName} (or {repo}) and {name} (or {branch}) placeholders. A / in the format creates nested folders.'
-                )}
-            </span>
-            <span className="block">
-              {translate(
-                'auto.components.settings.WorkspaceDirectorySetting.f6a7b8c9de',
-                'Example'
-              )}
-              : <code className="text-xs">{example}</code>
-            </span>
-          </>
-        }
-        control={
-          <Select value={mode} onValueChange={(v) => selectMode(v as WorktreeNamingMode)}>
-            <SelectTrigger id={`${inputId}-mode`} className="w-[180px]">
-              <SelectValue />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="flat">
-                {translate('auto.components.settings.WorkspaceDirectorySetting.9a0b1c2dbe', 'Flat')}
-              </SelectItem>
-              <SelectItem value="nested">
-                {translate(
-                  'auto.components.settings.WorkspaceDirectorySetting.a1b2c3d4ef',
-                  'Nested by repo'
-                )}
-              </SelectItem>
-              <SelectItem value="custom">
-                {translate(
-                  'auto.components.settings.WorkspaceDirectorySetting.b2c3d4e5fa',
-                  'Custom format…'
-                )}
-              </SelectItem>
-            </SelectContent>
-          </Select>
-        }
-      />
-      {mode === 'custom' && (
-        <Input
-          id={`${inputId}-format`}
-          value={customFormat}
-          placeholder="{repoName}.{name}"
-          onChange={(e) => setCustomFormat(e.target.value)}
-          className="text-xs"
-        />
-      )}
-    </div>
-  )
-}
-
-// Why: render an illustrative example so users see how their format resolves
-// before creating a worktree. Uses a sample repo/worktree name; the real path
-// is computed at creation time by computeWorktreePath in the main process.
-function renderWorktreeNameExample(
-  mode: WorktreeNamingMode,
-  customFormat: string,
-  workspaceDir: string
-): string {
-  const repoName = 'my-project'
-  const name = 'fix-bug'
-  let relative: string
-  if (mode === 'flat') {
-    relative = name
-  } else if (mode === 'nested') {
-    relative = `${repoName}/${name}`
-  } else {
-    relative =
-      customFormat
-        .replace(/\{repoName\}/g, repoName)
-        .replace(/\{repo\}/g, repoName)
-        .replace(/\{name\}/g, name)
-        .replace(/\{branch\}/g, name) || name
-  }
-  return `${workspaceDir.replace(/\/$/, '')}/${relative}`
 }
